requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
django-storages==1.14.4
boto3==1.35.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
pillow==11.1.0
python-barcode==0.15.1
qrcode==8.0
openpyxl==3.1.5
reportlab==4.2.5
anthropic>=0.42.0
requests>=2.31.0
PySocks>=1.7.1
pytest>=9.0
pytest-django>=4.11
